Milestone 2/requirements.txt
numpy
pandas
scikit-learn
PyYAML
requests
SQLAlchemy
psycopg2-binary
pyarrow
docker
kafka-python